requirements.txt
absl-py==1.4.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
attrs==22.2.0
autopep8==2.0.1
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.2.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
dnspython==2.3.0
emojis==0.7.0
exceptiongroup==1.1.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
flake8==6.0.0
flatbuffers==23.5.9
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
html5lib==1.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.10.1
importlib-metadata==7.0.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jax==0.4.10
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.41.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.1.0
MouseInfo==0.1.3
multimethod==1.10
multitasking==0.0.11
mypy==1.2.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.6.0
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
phik==0.12.4
Pillow==9.4.0
platformdirs==3.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.14.0
pymongo==4.3.3
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.9.3.3
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
pyqt6-tools==6.4.2.3.3
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytweening==1.0.4
pytz==2022.7.1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
qtconsole==5.5.1
QtPy==2.4.1
radian==0.6.4
rchitect==0.3.40
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.32.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.22.0
trio-websocket==0.10.2
typeguard==4.1.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.15
validators==0.22.0
virtualenv==20.20.0
virtualenvwrapper-win==1.2.7
visions==0.7.5
watchdog==4.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.4
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.2.0
ydata-profiling==4.6.4
yfinance==0.2.18
zipp==3.17.0
